feat: Add GitHub Actions workflow for Android release APK build
This workflow automates the process of building a release APK for the Android platform
from the 'client-lsf' branch. It includes steps for setting up the environment,
installing dependencies, running Expo prebuild, and executing the Gradle build
to generate the 'app-release.apk'. The generated APK is then uploaded as a workflow artifact.

The workflow is triggered on pushes to the 'client-lsf' branch and can also be
triggered manually via workflow_dispatch.